requirements.txt
openai==0.27.4
pinecone-client==2.2.1
python-dotenv
deeplake
Gunicorn
Flask
requests
requests_oauthlib
tweepy
langchain==0.0.144
redis
numpy
pandas
gdown
faiss-cpu
google-search-results
typing
tiktoken
pytz
wikipedia
re
schedule
